additional_requirements.txt
pyg-lib==0.1.0
torch_cluster==1.6.0
torch-spline-conv==1.2.1
torch_scatter==2.1.0
torch_sparse==0.6.15
torch_geometric==2.2.0